feat: 데이터 비교 유틸함수 생성 #109
